wrote compute_jacobian. it seems inefficient and buggy - look at it again later

+++ cpp/add_observation_noise.cpp
@@ -4,7 +4,7 @@
 void add_observation_noise(MatrixXf &z, MatrixXf R, int addnoise)
 {
 	if (addnoise == 1){
-		float len = z.cols();	
+		int len = z.cols();	
 		if (len > 0) {
 			MatrixXf randM1(1,len);
 			MatrixXf randM2(1,len);
